requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Flask-Migrate==4.0.0
Flask-HTTPAuth==4.7.0
Flask-Caching==2.1.0
cryptography==41.0.3
argon2-cffi==23.2.0
